feat: Update product data and enhance frontend UI/UX
- Replaced generic product data with realistic names, prices, and descriptions
- Added Graph-based product recommendations system with backend API
- Redesigned cart sidebar (left-aligned, improved typography, reduced overlay opacity)
- Implemented footer with subscription CTA and navigation sections
- Updated styling: Inter font, rectangular buttons, cursor-pointer on interactives
- Fixed checkout bug with Unix timestamp orderId (int → long long in C++)
- Enhanced visual consistency with cream accent color (#F2F0E6)

+++ cpp/src/order.cpp
@@ -23,7 +23,7 @@
 class Order
 {
 public:
-    int orderId;
+    long long orderId;
     int userId;
     std::vector<int> productIds;
     double totalPrice;
@@ -184,7 +184,7 @@ public:
             while (std::getline(ss, token, ','))
             {
                 if (fieldIndex == 0)
-                    newOrder.orderId = std::stoi(token);
+                    newOrder.orderId = std::stoll(token);
                 else if (fieldIndex == 1)
                     newOrder.userId = std::stoi(token);
                 else if (fieldIndex == 2)
